% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetClass.R
\name{assetClass}
\alias{assetClass}
\title{Practice Foundational Data Science Skills}
\usage{
assetClass(name = "assetClass")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Practice skills and presentation with a real life context problem.
}
\examples{
if (interactive()) {
  assetClass()
}
}
\author{
Philippe Cote
}
