% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetClasses.R
\name{assetClasses}
\alias{assetClasses}
\title{Produce a clean analysis of returns properties and correlation across asset classes.}
\usage{
assetClasses(name = "asset-classes")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Analyzing financial returns across multiple asset classes.
Your task is to perform analysis and render it as effective business communication.
}
\author{
Philippe Cote
}
