% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataWrangling.R
\name{dataWrangling}
\alias{dataWrangling}
\title{Data Wrangling Quiz}
\usage{
dataWrangling(name = "data-wrangling")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Testing your understanding of basic data wrangling concepts.
}
\examples{
dataWrangling()
}
\author{
Philippe Cote
}
