% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness}
\alias{fitness}
\title{Produce a recommendation for an emerging old athlete as a coach}
\usage{
fitness(name = "fitness")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Assignment using data from Strava user data.
}
\examples{
if (interactive()) {
  fitness()
}
}
\author{
Philippe Cote
}
