% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettingStarted.R
\name{gettingStarted}
\alias{gettingStarted}
\title{Getting Started Quiz}
\usage{
gettingStarted(name = "getting-started")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Testing your understanding of R concepts.
}
\examples{
\dontrun{
gettingStarted()
}
}
\author{
Philippe Cote
}
