% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketMaking.R
\name{marketMaking}
\alias{marketMaking}
\title{Energy Futures and Swaps Market Making - WIP}
\usage{
marketMaking(name = "marketMaking")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Once you understand how to price swaps from futures and
want to practice being a market market and hedging the risk this is for you.
GAME IS CURRENTLY IN DEVELOPMENT AND EXPERIMENTAL
}
\examples{
if (interactive()) {
  marketMaking()
}
}
\author{
Philippe Cote
}
