% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtmPLexpo.R
\name{mtmPLexpo}
\alias{mtmPLexpo}
\title{Compute PL and Exposure}
\usage{
mtmPLexpo(name = "mtmPLexpo")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Given a set a transaction you must compute Profit and Loss and current Exposure
}
\examples{
if (interactive()) {
  mtmPLexpo()
}
}
\author{
Philippe Cote
}
