% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleRegs.R
\name{multipleRegs}
\alias{multipleRegs}
\title{Produce a clean analysis of recursively running a multivariate regression.}
\usage{
multipleRegs(name = "multivariate-regressions")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Apply a practitioner's mindset to decide what to include and excludes in your multiple regression.
Your task is to perform analysis and render it as effective business communication.
}
\examples{
multipleRegs()
}
\author{
Philippe Cote
}
