% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{orders}
\alias{orders}
\title{Execution orders}
\usage{
orders(name = "orders")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Practice your understanding of market and limit orders.
}
\examples{
orders()
}
\author{
Philippe Cote
}
