% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskIntoPractice.R
\name{riskIntoPractice}
\alias{riskIntoPractice}
\title{Practice Foundational Risk Management Skills}
\usage{
riskIntoPractice(name = "riskIntoPractice")
}
\arguments{
\item{name}{riskIntoPractice}
}
\value{
'learnr' module
}
\description{
Practice RM skills and presentation with a real life context problem.
}
\examples{
if (interactive()) {
  riskIntoPractice()
}
}
\author{
Philippe Cote
}
