% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskintopractice-energy.R
\name{riskintopracticeEnergy}
\alias{riskintopracticeEnergy}
\title{Practice Fondational Risk Management Skills}
\usage{
riskintopracticeEnergy(name = "riskintopractice-energy")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Practice RM skills and presentation with a real life context problem.
}
\examples{
if (interactive()) {
  riskintopracticeEnergy()
}
}
\author{
Philippe Cote
}
