% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmCanCrude.R
\name{rmCanCrude}
\alias{rmCanCrude}
\title{Practice Foundational Risk Management Skills}
\usage{
rmCanCrude(name = "rmCanCrude")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Practice RM skills and presentation with a real life context problem.
}
\examples{
if (interactive()) {
  rmCanCrude()
}
}
\author{
Philippe Cote
}
