% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmNYHulsd.R
\name{rmNYHulsd}
\alias{rmNYHulsd}
\title{Practice Foundational Risk Management Skills}
\usage{
rmNYHulsd(name = "rmNYHulsd")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Practice RM skills and presentation with a real life context problem.
}
\examples{
if (interactive()) {
  rmNYHulsd()
}
}
\author{
Philippe Cote
}
