% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTLedu-package.R
\docType{package}
\name{RTLedu-package}
\alias{RTLedu}
\alias{RTLedu-package}
\title{RTLedu: 'Learnr' Modules in Finance}
\description{
Build to support the 'RTL' package in education and market practitioners learning R, Finance and Commodities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/risktoollib/RTLedu}
}

}
\author{
\strong{Maintainer}: Philippe Cote \email{pcote@ualberta.ca}

}
\keyword{internal}
