% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp400_prices}
\alias{sp400_prices}
\title{sp400_prices data}
\format{
data frame
}
\usage{
sp400_prices
}
\description{
S&P500 components monthly adjusted stock price returns
}
\keyword{datasets}
