% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot2fut.R
\name{spot2fut}
\alias{spot2fut}
\title{Spot To Futures Relationship}
\usage{
spot2fut(name = "spot2fut")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Deepen and test our understanding of the important spot to futures relationship in Finance.
}
\examples{
\dontrun{
spot2fut()
}
}
\author{
Philippe Cote
}
