% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spreads}
\alias{spreads}
\title{Sample chart for spreads trading}
\format{
list
}
\usage{
spreads
}
\description{
Effective understanding of risk of specific contract pairs
}
\keyword{datasets}
