% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usGeneration.R
\name{usGeneration}
\alias{usGeneration}
\title{Produce a clean analysis on the US Electricity Market}
\usage{
usGeneration(name = "usGeneration")
}
\arguments{
\item{name}{Tutorial name}
}
\value{
'learnr' module
}
\description{
Assignment using data from the Energy Information Administration (EIA) API.
}
\examples{
if (interactive()) {
  usGeneration()
}
}
\author{
Philippe Cote
}
